🔨 refactor(app): add ip service
